Flesh out detail page, adding TOC and charts
Add table of contents and the three overall sections (Publications,
Patents, and Workforce) to the company detail page.  Add the initial
set of charts, including helper functions to standardize things as
much as possible.

The current version of the URL slug -> company ID/name extraction
won't work when used with Gatsby's `<Link>` component, since it only
runs once on the initial page load (so navigating to a second company
would still present the first company on the page).