Fix bug where incorrect initial columns were selected
Update the initial selections of columns to properly reflect the
already-selected columns on first load of the Add/Remove Columns
dialog.

Previously, adjusting columns and then reloading the page (or loading
from the start with `zz_columns` set) would lead to the Add/Remove
Column dialog loading with _all_ columns selected (the defaults), not
the currently specified columns.

Closes #29