Hide custom group membership when a different group is selected
When the user switches from a custom group to a pre-defined group,
remove the group's membership list from URL and store it in a separate
state variable.  This allows users to share filter URLs without also
unintentionally sharing their custom groups (they'd have to be in the
custom group mode in order to share that info).