Fix persistence issue with header dropdowns
Resolve issue where the default value for a dropdown filter was being
passed to the underlying `<Autocomplete>` before the value retrieved
from the URL parameters could be provided, leading to the Autocomplete
overwriting the desired value with the default value.  To fix this,
update the dropdown filters with the initial filter state in a
first-render `useEffect` call, which occurs after the unwanted
overwriting but before any user interactions could occur.

Closes #179