Change aggregation to separate parent/child permids
Note: test_aggregate_organizations.py won't quite work right now until the legacy CSET ids fix is implemented because there are hard-coded legacy CSET ids that aren't getting non-aggregated correctly. But I tested it with them switched over to the new CSET ids and it works, so once that fix is implemented it should be fine.

Also, I haven't implemented use of the parent permid for linking to a sector; I'll leave that to you.