Rework buttonbar label visibility based on viewport
Adjust the visibility of the labels for the Reset, Download, and Column
buttons based on the available space in the buttonbar.  Download's label
will hide first (still accessible to screen readers), followed by
Add/Remove Columns, and finally Reset.  If not enough space is available
for the "Viewing X of Y..." text and/or buttons, they will wrap onto the
next row of the buttonbar.

Closes #464